Refactor event filtering into eventbridge utils
